+++ cmd/psqldef/psqldef_test.go
@@ -568,9 +568,38 @@ func TestPsqldefCreateView(t *testing.T) {
 	}
 }
 
+func TestPsqldefCreateMaterializedView(t *testing.T) {
+	for _, tc := range publicAndNonPublicSchemaTestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			resetTestDatabase()
+			mustExecuteSQL(fmt.Sprintf("CREATE SCHEMA IF NOT EXISTS %s;", tc.Schema))
+
+			createUsers := fmt.Sprintf("CREATE TABLE %s.users (id BIGINT PRIMARY KEY, name character varying(100));\n", tc.Schema)
+			createPosts := fmt.Sprintf("CREATE TABLE %s.posts (id BIGINT PRIMARY KEY, name character varying(100), user_id BIGINT, is_deleted boolean);\n", tc.Schema)
+			assertApplyOutput(t, createUsers+createPosts, applyPrefix+createUsers+createPosts)
+			assertApplyOutput(t, createUsers+createPosts, nothingModified)
+
+			posts := "posts"
+			users := "users"
+			if tc.Schema != "public" {
+				posts = fmt.Sprintf("%s.posts", tc.Schema)
+				users = fmt.Sprintf("%s.users", tc.Schema)
+			}
+
+			createMaterializedView := fmt.Sprintf("CREATE MATERIALIZED VIEW %s.view_user_posts AS SELECT p.id FROM (%s as p JOIN %s as u ON ((p.user_id = u.id)));\n", tc.Schema, posts, users)
+			assertApplyOutput(t, createUsers+createPosts+createMaterializedView, applyPrefix+
+				fmt.Sprintf("CREATE MATERIALIZED VIEW %s.view_user_posts AS SELECT p.id FROM (%s as p JOIN %s as u ON ((p.user_id = u.id)));\n", tc.Schema, posts, users))
+			assertApplyOutput(t, createUsers+createPosts+createMaterializedView, nothingModified)
+		})
+	}
+}
+
 func TestPsqldefDropPrimaryKey(t *testing.T) {
 	resetTestDatabase()
+<<<<<<< HEAD
 
+=======
+>>>>>>> 1d3e1f675ed0dc90e232d35a936ac5163e5d3df4
 	createTable := stripHeredoc(`
 		CREATE TABLE users (
 		  id bigint NOT NULL PRIMARY KEY,
@@ -627,6 +656,38 @@ func TestPsqldefCreateIndex(t *testing.T) {
 				dropIndex2+"\n"+
 				dropIndex1+"\n")
 			assertApplyOutput(t, createTable, nothingModified)
+		})
+	}
+}
+
+func TestPsqldefCreateMaterializedViewIndex(t *testing.T) {
+	for _, tc := range publicAndNonPublicSchemaTestCases {
+		t.Run(tc.Name, func(t *testing.T) {
+			resetTestDatabase()
+			mustExecuteSQL(fmt.Sprintf("CREATE SCHEMA IF NOT EXISTS %s;", tc.Schema))
+
+			createTable := stripHeredoc(fmt.Sprintf(`
+				CREATE TABLE %s.users (
+				  id bigint NOT NULL,
+				  name text,
+				  age integer
+				);`, tc.Schema))
+			users := "users"
+			if tc.Schema != "public" {
+				users = fmt.Sprintf("%s.users", tc.Schema)
+			}
+			createMaterializedView := fmt.Sprintf("CREATE MATERIALIZED VIEW %s.view_users AS SELECT * FROM %s;\n", tc.Schema, users)
+			assertApplyOutput(t, createTable+createMaterializedView, applyPrefix+
+				createTable+"\n"+
+				fmt.Sprintf("CREATE MATERIALIZED VIEW %s.view_users AS SELECT * FROM %s;\n", tc.Schema, users))
+			assertApplyOutput(t, createTable+createMaterializedView, nothingModified)
+
+			createIndex1 := fmt.Sprintf(`CREATE INDEX index_name on %s.view_users (name);`, tc.Schema)
+			createIndex2 := fmt.Sprintf(`CREATE UNIQUE INDEX index_age on %s.view_users (age);`, tc.Schema)
+			assertApplyOutput(t, createTable+createMaterializedView+createIndex1+createIndex2, applyPrefix+
+				createIndex1+"\n"+
+				createIndex2+"\n")
+			assertApplyOutput(t, createTable+createMaterializedView+createIndex1+createIndex2, nothingModified)
 		})
 	}
 }
@@ -1092,7 +1153,7 @@ func TestPsqldefAddUniqueConstraintToTableInNonpublicSchema(t *testing.T) {
 		    "b" integer
 		);
 
-		ALTER TABLE test.dummy ADD CONSTRAINT a_b_uniq UNIQUE (a, b);
+		ALTER TABLE "test"."dummy" ADD CONSTRAINT "a_b_uniq" UNIQUE (a, b);
 		`))
 	assertApplyOutput(t, createTable+"\n"+alterTable, nothingModified)
 
@@ -1106,7 +1167,7 @@ func TestPsqldefAddUniqueConstraintToTableInNonpublicSchema(t *testing.T) {
 		    "b" integer
 		);
 
-		ALTER TABLE test.dummy ADD CONSTRAINT a_uniq UNIQUE (a) DEFERRABLE INITIALLY DEFERRED;
+		ALTER TABLE "test"."dummy" ADD CONSTRAINT "a_uniq" UNIQUE (a) DEFERRABLE INITIALLY DEFERRED;
 		`))
 	assertApplyOutput(t, createTable+"\n"+alterTable, nothingModified)
 }
@@ -1221,7 +1282,7 @@ func TestPsqldefExport(t *testing.T) {
 		    PRIMARY KEY ("id")
 		);
 
-		ALTER TABLE public.users ADD CONSTRAINT users_c_char_1_key UNIQUE (c_char_1);
+		ALTER TABLE "public"."users" ADD CONSTRAINT "users_c_char_1_key" UNIQUE (c_char_1);
 		`,
 	))
 }