+++ database/postgres/database.go
@@ -171,7 +171,9 @@ func (d *PostgresDatabase) materializedViews() ([]string, error) {
 		if err != nil {
 			return ddls, err
 		}
-		ddls = append(ddls, indexDefs...)
+		for _, indexDef := range indexDefs {
+			ddls = append(ddls, fmt.Sprintf("%s;", indexDef))
+		}
 	}
 	return ddls, nil
 }
@@ -574,7 +576,11 @@ func (d *PostgresDatabase) getUniqueConstraints(tableName string) (map[string]st
 		if err != nil {
 			return nil, err
 		}
-		result[constraintName] = fmt.Sprintf("ALTER TABLE %s ADD CONSTRAINT %s %s", tableName, constraintName, constraintDef)
+
+		result[constraintName] = fmt.Sprintf("ALTER TABLE %s.%s ADD CONSTRAINT %s %s",
+			escapeSQLName(schema), escapeSQLName(table),
+			escapeSQLName(constraintName), constraintDef,
+		)
 	}
 
 	return result, nil